// src/generated/graphql.ts
export type Maybe<T> = T | null;
// ...
export type CodeOfConduct = Node & {
  __typename?: 'CodeOfConduct';
  /** The body of the Code of Conduct */
  body?: Maybe<Scalars['String']>;
  id: Scalars['ID'];
  /** The key for the Code of Conduct */
  key: Scalars['String'];
  /** The formal name of the Code of Conduct */
  name: Scalars['String'];
  /** The HTTP path for this Code of Conduct */
  resourcePath?: Maybe<Scalars['URI']>;
  /** The HTTP URL for this Code of Conduct */
  url?: Maybe<Scalars['URI']>;
};
// ...
export type Node = {
  /** ID of the object. */
  id: Scalars['ID'];
};
// ...
export type Query = {
  __typename?: 'Query';
  /** Look up a code of conduct by its key */
  codeOfConduct?: Maybe<CodeOfConduct>;
  /** Look up a code of conduct by its key */
  codesOfConduct?: Maybe<Array<Maybe<CodeOfConduct>>>;
  /** Look up an enterprise by URL slug. */
  enterprise?: Maybe<Enterprise>;
  /** Look up a pending enterprise administrator invitation by invitee, enterprise and role. */
  enterpriseAdministratorInvitation?: Maybe<EnterpriseAdministratorInvitation>;
  /** Look up a pending enterprise administrator invitation by invitation token. */
  enterpriseAdministratorInvitationByToken?: Maybe<EnterpriseAdministratorInvitation>;
  /** Look up an open source license by its key */
  license?: Maybe<License>;
  /** Return a list of known open source licenses */
  licenses: Array<Maybe<License>>;
  /** Get alphabetically sorted list of Marketplace categories */
  marketplaceCategories: Array<MarketplaceCategory>;
  /** Look up a Marketplace category by its slug. */
  marketplaceCategory?: Maybe<MarketplaceCategory>;
  /** Look up a single Marketplace listing */
  marketplaceListing?: Maybe<MarketplaceListing>;
  /** Look up Marketplace listings */
  marketplaceListings: MarketplaceListingConnection;
  /** Return information about the GitHub instance */
  meta: GitHubMetadata;
  /** Fetches an object given its ID. */
  node?: Maybe<Node>;
  /** Lookup nodes by a list of IDs. */
  nodes: Array<Maybe<Node>>;
  /** Lookup a organization by login. */
  organization?: Maybe<Organization>;
  /** The client's rate limit information. */
  rateLimit?: Maybe<RateLimit>;
  /** Hack to workaround https://github.com/facebook/relay/issues/112 re-exposing the root query object */
  relay: Query;
  /** Lookup a given repository by the owner and repository name. */
  repository?: Maybe<Repository>;
  /** Lookup a repository owner (ie. either a User or an Organization) by login. */
  repositoryOwner?: Maybe<RepositoryOwner>;
  /** Lookup resource by a URL. */
  resource?: Maybe<UniformResourceLocatable>;
  /** Perform a search across resources. */
  search: SearchResultItemConnection;
  /** GitHub Security Advisories */
  securityAdvisories: SecurityAdvisoryConnection;
  /** Fetch a Security Advisory by its GHSA ID */
  securityAdvisory?: Maybe<SecurityAdvisory>;
  /** Software Vulnerabilities documented by GitHub Security Advisories */
  securityVulnerabilities: SecurityVulnerabilityConnection;
  /**
   * Look up a single Sponsors Listing
   * @deprecated `Query.sponsorsListing` will be removed. Use `Sponsorable.sponsorsListing` instead. Removal on 2020-04-01 UTC.
   */
  sponsorsListing?: Maybe<SponsorsListing>;
  /** Look up a topic by name. */
  topic?: Maybe<Topic>;
  /** Lookup a user by login. */
  user?: Maybe<User>;
  /** The currently authenticated user. */
  viewer: User;
};
// ...
export type UniformResourceLocatable = {
  /** The HTML path to this resource. */
  resourcePath: Scalars['URI'];
  /** The URL to this resource. */
  url: Scalars['URI'];
};
// ...
export type CreateIssueMutationVariables = Exact<{
  body?: Maybe<Scalars['String']>;
  labelIds?: Maybe<Array<Scalars['ID']>>;
  repositoryId: Scalars['ID'];
  title: Scalars['String'];
}>;


export type CreateIssueMutation = (
  { __typename?: 'Mutation' }
  & { createIssue?: Maybe<(
    { __typename?: 'CreateIssuePayload' }
    & Pick<CreateIssuePayload, 'clientMutationId'>
  )> }
);

export type CreateLabelMutationVariables = Exact<{
  repositoryId: Scalars['ID'];
  name: Scalars['String'];
  color: Scalars['String'];
  description?: Maybe<Scalars['String']>;
}>;


export type CreateLabelMutation = (
  { __typename?: 'Mutation' }
  & { createLabel?: Maybe<(
    { __typename?: 'CreateLabelPayload' }
    & Pick<CreateLabelPayload, 'clientMutationId'>
  )> }
);

export type CreateRepositoryMutationVariables = Exact<{
  ownerId?: Maybe<Scalars['ID']>;
  name: Scalars['String'];
}>;


export type CreateRepositoryMutation = (
  { __typename?: 'Mutation' }
  & { createRepository?: Maybe<(
    { __typename?: 'CreateRepositoryPayload' }
    & { repository?: Maybe<(
      { __typename?: 'Repository' }
      & Pick<Repository, 'id' | 'isTemplate' | 'url'>
    )> }
  )> }
);

export type DeleteLabelMutationVariables = Exact<{
  id: Scalars['ID'];
}>;


export type DeleteLabelMutation = (
  { __typename?: 'Mutation' }
  & { deleteLabel?: Maybe<(
    { __typename?: 'DeleteLabelPayload' }
    & Pick<DeleteLabelPayload, 'clientMutationId'>
  )> }
);

export type GetIssuesQueryVariables = Exact<{
  owner: Scalars['String'];
  name: Scalars['String'];
  cursor?: Maybe<Scalars['String']>;
}>;


export type GetIssuesQuery = (
  { __typename?: 'Query' }
  & { repository?: Maybe<(
    { __typename?: 'Repository' }
    & Pick<Repository, 'id'>
    & { issues: (
      { __typename?: 'IssueConnection' }
      & { pageInfo: (
        { __typename?: 'PageInfo' }
        & Pick<PageInfo, 'endCursor' | 'hasNextPage'>
      ), edges?: Maybe<Array<Maybe<(
        { __typename?: 'IssueEdge' }
        & { node?: Maybe<(
          { __typename?: 'Issue' }
          & Pick<Issue, 'title' | 'body'>
          & { labels?: Maybe<(
            { __typename?: 'LabelConnection' }
            & { nodes?: Maybe<Array<Maybe<(
              { __typename?: 'Label' }
              & Pick<Label, 'name'>
            )>>> }
          )> }
        )> }
      )>>> }
    ) }
  )> }
);

export type GetLabelsQueryVariables = Exact<{
  owner: Scalars['String'];
  name: Scalars['String'];
  cursor?: Maybe<Scalars['String']>;
}>;


export type GetLabelsQuery = (
  { __typename?: 'Query' }
  & { repository?: Maybe<(
    { __typename?: 'Repository' }
    & Pick<Repository, 'id'>
    & { labels?: Maybe<(
      { __typename?: 'LabelConnection' }
      & { pageInfo: (
        { __typename?: 'PageInfo' }
        & Pick<PageInfo, 'endCursor' | 'hasNextPage'>
      ), edges?: Maybe<Array<Maybe<(
        { __typename?: 'LabelEdge' }
        & { node?: Maybe<(
          { __typename?: 'Label' }
          & Pick<Label, 'id' | 'color' | 'name' | 'description'>
        )> }
      )>>> }
    )> }
  )> }
);

export type GetRepositoryQueryVariables = Exact<{
  owner: Scalars['String'];
  name: Scalars['String'];
}>;


export type GetRepositoryQuery = (
  { __typename?: 'Query' }
  & { repository?: Maybe<(
    { __typename?: 'Repository' }
    & Pick<Repository, 'id' | 'url' | 'isTemplate'>
  )> }
);

export type GetUserQueryVariables = Exact<{
  login: Scalars['String'];
}>;


export type GetUserQuery = (
  { __typename?: 'Query' }
  & { user?: Maybe<(
    { __typename?: 'User' }
    & Pick<User, 'id'>
  )> }
);

export type ImportRepositoryMutationVariables = Exact<{
  name: Scalars['String'];
  ownerId: Scalars['ID'];
  repositoryId: Scalars['ID'];
}>;


export type ImportRepositoryMutation = (
  { __typename?: 'Mutation' }
  & { cloneTemplateRepository?: Maybe<(
    { __typename?: 'CloneTemplateRepositoryPayload' }
    & { repository?: Maybe<(
      { __typename?: 'Repository' }
      & Pick<Repository, 'id' | 'url' | 'isTemplate'>
    )> }
  )> }
);

export const CreateIssueDocument = gql`
    mutation createIssue($body: String, $labelIds: [ID!], $repositoryId: ID!, $title: String!) {
  createIssue(input: {body: $body, labelIds: $labelIds, repositoryId: $repositoryId, title: $title}) {
    clientMutationId
  }
}
    `;

  @Injectable({
    providedIn: 'root'
  })
  export class CreateIssueGQL extends Apollo.Mutation<CreateIssueMutation, CreateIssueMutationVariables> {
    document = CreateIssueDocument;
    
    constructor(apollo: Apollo.Apollo) {
      super(apollo);
    }
  }
export const CreateLabelDocument = gql`
    mutation createLabel($repositoryId: ID!, $name: String!, $color: String!, $description: String) {
  createLabel(input: {repositoryId: $repositoryId, name: $name, color: $color, description: $description}) {
    clientMutationId
  }
}
    `;

  @Injectable({
    providedIn: 'root'
  })
  export class CreateLabelGQL extends Apollo.Mutation<CreateLabelMutation, CreateLabelMutationVariables> {
    document = CreateLabelDocument;
    
    constructor(apollo: Apollo.Apollo) {
      super(apollo);
    }
  }
export const CreateRepositoryDocument = gql`
    mutation createRepository($ownerId: ID, $name: String!) {
  createRepository(input: {ownerId: $ownerId, name: $name, visibility: PUBLIC}) {
    repository {
      id
      isTemplate
      url
    }
  }
}
    `;

  @Injectable({
    providedIn: 'root'
  })
  export class CreateRepositoryGQL extends Apollo.Mutation<CreateRepositoryMutation, CreateRepositoryMutationVariables> {
    document = CreateRepositoryDocument;
    
    constructor(apollo: Apollo.Apollo) {
      super(apollo);
    }
  }
export const DeleteLabelDocument = gql`
    mutation deleteLabel($id: ID!) {
  deleteLabel(input: {id: $id}) {
    clientMutationId
  }
}
    `;

  @Injectable({
    providedIn: 'root'
  })
  export class DeleteLabelGQL extends Apollo.Mutation<DeleteLabelMutation, DeleteLabelMutationVariables> {
    document = DeleteLabelDocument;
    
    constructor(apollo: Apollo.Apollo) {
      super(apollo);
    }
  }
export const GetIssuesDocument = gql`
    query getIssues($owner: String!, $name: String!, $cursor: String) {
  repository(name: $name, owner: $owner) {
    id
    issues(first: 100, states: OPEN, after: $cursor) {
      pageInfo {
        endCursor
        hasNextPage
      }
      edges {
        node {
          title
          labels(first: 100) {
            nodes {
              name
            }
          }
          body
        }
      }
    }
  }
}
    `;

  @Injectable({
    providedIn: 'root'
  })
  export class GetIssuesGQL extends Apollo.Query<GetIssuesQuery, GetIssuesQueryVariables> {
    document = GetIssuesDocument;
    
    constructor(apollo: Apollo.Apollo) {
      super(apollo);
    }
  }
export const GetLabelsDocument = gql`
    query getLabels($owner: String!, $name: String!, $cursor: String) {
  repository(name: $name, owner: $owner) {
    id
    labels(first: 100, after: $cursor) {
      pageInfo {
        endCursor
        hasNextPage
      }
      edges {
        node {
          id
          color
          name
          description
        }
      }
    }
  }
}
    `;

  @Injectable({
    providedIn: 'root'
  })
  export class GetLabelsGQL extends Apollo.Query<GetLabelsQuery, GetLabelsQueryVariables> {
    document = GetLabelsDocument;
    
    constructor(apollo: Apollo.Apollo) {
      super(apollo);
    }
  }
export const GetRepositoryDocument = gql`
    query getRepository($owner: String!, $name: String!) {
  repository(name: $name, owner: $owner) {
    id
    url
    isTemplate
  }
}
    `;

  @Injectable({
    providedIn: 'root'
  })
  export class GetRepositoryGQL extends Apollo.Query<GetRepositoryQuery, GetRepositoryQueryVariables> {
    document = GetRepositoryDocument;
    
    constructor(apollo: Apollo.Apollo) {
      super(apollo);
    }
  }
export const GetUserDocument = gql`
    query getUser($login: String!) {
  user(login: $login) {
    id
  }
}
    `;

  @Injectable({
    providedIn: 'root'
  })
  export class GetUserGQL extends Apollo.Query<GetUserQuery, GetUserQueryVariables> {
    document = GetUserDocument;
    
    constructor(apollo: Apollo.Apollo) {
      super(apollo);
    }
  }
export const ImportRepositoryDocument = gql`
    mutation importRepository($name: String!, $ownerId: ID!, $repositoryId: ID!) {
  cloneTemplateRepository(input: {name: $name, ownerId: $ownerId, repositoryId: $repositoryId, visibility: PUBLIC}) {
    repository {
      id
      url
      isTemplate
    }
  }
}
    `;

  @Injectable({
    providedIn: 'root'
  })
  export class ImportRepositoryGQL extends Apollo.Mutation<ImportRepositoryMutation, ImportRepositoryMutationVariables> {
    document = ImportRepositoryDocument;
    
    constructor(apollo: Apollo.Apollo) {
      super(apollo);
    }
  }